Major Breakthrough
I finally figured out how to make custom spells appear in the proper spell tab in the spellbook AND appear on the proper trainer. 

Adding them to the npc_trainer table didn't do the trick. You need to edit the SkillLineAbility.dbc to make it actually work. I was never able to find any useful hints towards this for the last few years I've tried on and off. I will probably upload a guide on ownedcore at somepoint.